requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
gunicorn==25.1.0
python-dotenv==1.2.2
qdrant-client==1.17.0
llama-index-core==0.14.15
llama-index-vector-stores-qdrant==0.9.2
llama-index-embeddings-huggingface==0.6.1
llama-index-llms-groq==0.4.1
sentence-transformers==5.2.3